requirements.txt
# core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
pydantic>=2.9.0

# embeddings + vector store
openai>=1.50.0
qdrant-client>=1.12.0

# document loaders
pypdf>=5.0.0

# cache
redis>=5.0.0

# observability
python-dotenv>=1.0.0
rich>=13.9.0

# api
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# http client (usado em examples/run_questions.py)
httpx>=0.27.0

# dev / test
pytest>=8.0.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
